% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islr_cnn_code.R
\name{build_array}
\alias{build_array}
\title{make an array of JPEG-like data for invocations of a (presumably random) shape-generating function}
\usage{
build_array(
  shapefun = rancirc,
  nimages = 2500,
  sidelength = 64,
  depth = 3,
  npts = 90,
  ...
)
}
\arguments{
\item{shapefun}{a function producing a sidelength x sidelength binary image}

\item{nimages}{numeric(1) number of images}

\item{sidelength}{numeric(1) passed to load_jpeg as siz}

\item{depth}{numeric(1) defaults to 3}

\item{npts}{numeric(1) passed to shapefun}
}
\description{
make an array of JPEG-like data for invocations of a (presumably random) shape-generating function
}
\examples{
circarr = build_array( rancirc, sidelength=32, side_plane=32, npts=120 )
dim(circarr)
}
