% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cifar100_history}
\alias{cifar100_history}
\title{a history object for keras fit for cifar100 data, using the ISLR tuning settings}
\format{
list
}
\usage{
cifar100_history
}
\description{
a history object for keras fit for cifar100 data, using the ISLR tuning settings
}
\note{
just saved the history() output for a keras fit
}
\keyword{datasets}
