% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ciftest2k}
\alias{ciftest2k}
\title{random collection of 2000 images from the test data of the cifar100 (100 categories) dataset}
\format{
ImageArray
}
\usage{
ciftest2k
}
\description{
random collection of 2000 images from the test data of the cifar100 (100 categories) dataset
}
\examples{
data(ciftest2k)
preview(ciftest2k) # will use keras
}
\keyword{datasets}
