% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ciftrain1k}
\alias{ciftrain1k}
\title{random collection of 1000 images from the training data of the cifar100 (100 categories) dataset}
\format{
ImageArray
}
\usage{
ciftrain1k
}
\description{
random collection of 1000 images from the training data of the cifar100 (100 categories) dataset
}
\examples{
data(ciftrain1k)
preview(ciftrain1k) # will use keras
}
\keyword{datasets}
