% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_accelerator.R
\name{configure_accelerator}
\alias{configure_accelerator}
\title{claude-recommended approach to agnostic config of GPU for mac and linux}
\usage{
configure_accelerator(verbose = TRUE)
}
\arguments{
\item{verbose}{logical}
}
\description{
claude-recommended approach to agnostic config of GPU for mac and linux
}
\note{
Proposition is that if tensorflow-metal is available, MPS will be found and used,
certainly by optimizer_adam() if used.
}
