% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islr_cnn_code.R
\name{eval_model}
\alias{eval_model}
\title{estimate accuracy of fitted model predictions for a given ImageArray}
\usage{
eval_model(model, iarr)
}
\arguments{
\item{iarr}{ImageArray instance}

\item{fitted}{islr_cnn model instance}
}
\value{
accuracy estimate
}
\description{
estimate accuracy of fitted model predictions for a given ImageArray
}
