% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageArray.R
\name{filterByType}
\alias{filterByType}
\title{return an ImageArray with images restricted to specific types}
\usage{
filterByType(iarr, tvec)
}
\arguments{
\item{iarr}{ImageArray instance}

\item{tvec}{character vector of types to retain}
}
\description{
return an ImageArray with images restricted to specific types
}
