% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageArray.R
\name{flattenToMatrix}
\alias{flattenToMatrix}
\title{linearize data in an image to a single vector for each image in an 
ImageArray instance}
\usage{
flattenToMatrix(iarr)
}
\arguments{
\item{iarr}{ImageArray instance}
}
\description{
linearize data in an image to a single vector for each image in an 
ImageArray instance
}
