% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fitted_cifar100.R
\name{get_fitted_cifar100}
\alias{get_fitted_cifar100}
\title{restore a fitted model for CIFAR100}
\usage{
get_fitted_cifar100()
}
\value{
a keras model reference
}
\description{
restore a fitted model for CIFAR100
}
\note{
Metadata about the input data are available in extdata/cif.json.
}
\examples{
mod = get_fitted_cifar100()
mod
}
