% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islr_cnn_code.R
\name{get_k_argmax}
\alias{get_k_argmax}
\title{produce discrete predictions with fitted model predictions for a given ImageArray}
\usage{
get_k_argmax(model, iarr)
}
\arguments{
\item{iarr}{ImageArray instance}

\item{fitted}{islr_cnn model instance}
}
\description{
produce discrete predictions with fitted model predictions for a given ImageArray
}
