% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getApple.R
\name{getApple}
\alias{getApple}
\title{get a high-resolution jpeg of an apple in ImageArray}
\usage{
getApple()
}
\description{
get a high-resolution jpeg of an apple in ImageArray
}
\examples{
ap1 = getApple()
ap1
plotOne(ap1)
}
