% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketri.R
\name{getline}
\alias{getline}
\title{produce sequence of points on a line between two points}
\usage{
getline(x1, y1, x2, y2, npts = 20)
}
\arguments{
\item{x1}{numeric}

\item{y1}{numeric}

\item{x2}{numeric}

\item{y2}{numeric}

\item{npts}{number of points to return}
}
\value{
list with elements x, y
}
\description{
produce sequence of points on a line between two points
}
\examples{
getline(0,0,3,5)
}
