% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageArray.R
\docType{class}
\name{ImageArray-class}
\alias{ImageArray-class}
\title{container for 4D image array (index,x,y,z), typically (,32,32,3)}
\description{
container for 4D image array (index,x,y,z), typically (,32,32,3)
}
