% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islr_cnn_code.R
\name{islr_cnn}
\alias{islr_cnn}
\title{use the CNN in Introduction to Statistical Learning with R on an ImageArray instance}
\usage{
islr_cnn(iarr, nEpochs = 30, batchSize = 128, valSplit = 0.2)
}
\arguments{
\item{iarr}{ImageArray instance, must have 32x32x3 images}

\item{nEpochs}{numeric(1) used in call to fit for keras_model_sequential}

\item{batchSize}{numeric(1) used in call to fit for keras_model_sequential}

\item{valSplit}{numeric(1) used in call to fit for keras_model_sequential}
}
\value{
list with components model (compiled model), history (of fit), and
vector of unique type labels used for inputs, marked as an S3 class
islr_cnn
}
\description{
use the CNN in Introduction to Statistical Learning with R on an ImageArray instance
}
