% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpeg2iarr.R
\name{jpeg2iarr}
\alias{jpeg2iarr}
\title{make an ImageArray with a single jpeg}
\usage{
jpeg2iarr(fn, type, typelevels)
}
\arguments{
\item{fn}{path to jpeg}

\item{type}{typelevels character()}
}
\description{
make an ImageArray with a single jpeg
}
\examples{
purp = system.file("jpegs", "newpurple.jpg", package="littleDeep")
c3 = restore_islr_cnn(system.file("extdata", "cif3", package="littleDeep")) 
ia = jpeg2iarr(purp, "rose", c3$typelevels) 
scores_string(c3$model, ia)
}
