% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getApple.R
\name{jpg2ImageArray}
\alias{jpg2ImageArray}
\title{produce an ImageArray for any single jpeg}
\usage{
jpg2ImageArray(jpgpath, type = "type not set")
}
\arguments{
\item{jpgpath}{character(1) path}

\item{type}{character(1) type label}
}
\description{
produce an ImageArray for any single jpeg
}
\examples{
ia = jpg2ImageArray(applePath())
plotOne(ia) # slow
}
