% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketri.R
\name{load_jpeg}
\alias{load_jpeg}
\title{this function defines a 3-dimensional array with given dimensions
and sets all elements to 1 except those with element indices given by x and y,
which are set to zero}
\usage{
load_jpeg(x, y, siz = 64, dim = c(64, 64, 3))
}
\arguments{
\item{x}{indices in first dimension of array}

\item{y}{indices in second dimension of array}

\item{siz}{numeric side length}

\item{dim}{numeric(3) dimensions of array to be filled}
}
\description{
this function defines a 3-dimensional array with given dimensions
and sets all elements to 1 except those with element indices given by x and y,
which are set to zero
}
\examples{
ntri = maketri()
jp = load_jpeg(ntri$x, ntri$y)
plot(0,0,type="n",xlim=c(0,1), ylim=c(0,1))
graphics::rasterImage(jp, 0, 0, 1, 1, interpolate=FALSE)
}
