% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islr_cnn_code.R
\name{make_shape_iarr}
\alias{make_shape_iarr}
\title{produce an ImageArray of circles, triangles, quadrilaterals}
\usage{
make_shape_iarr(nimages = 2500)
}
\arguments{
\item{nimages}{numeric(1)}
}
\description{
produce an ImageArray of circles, triangles, quadrilaterals
}
