% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makecirc.R
\name{makecirc}
\alias{makecirc}
\title{make a discrete image of a circle}
\usage{
makecirc(center_x = 30, center_y = 30, radius = 15, npts = 100)
}
\arguments{
\item{center_x}{numeric(1) defaults to 30}

\item{center_y}{numeric(1) defaults to 30}

\item{radius}{numeric(1) defaults to 15}

\item{npts}{numeric(1) number of points to generate using `seq(0, 2*pi, length.out=npts)`}
}
\description{
make a discrete image of a circle
}
\examples{
plot(makecirc())
}
