% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEBImage.R
\name{makeEBImage}
\alias{makeEBImage}
\title{convert a length-1 ImageArray to an Image object in EBImage, via jpeg export}
\usage{
makeEBImage(iarr1)
}
\arguments{
\item{iarr1}{ImageArray instance}
}
\description{
convert a length-1 ImageArray to an Image object in EBImage, via jpeg export
}
\examples{
data(ciftrain1k)
ore = makeEBImage(ciftrain1k[4])
EBImage::display(ore, method="raster") # small!
}
