% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makequad.R
\name{makequad}
\alias{makequad}
\title{produce a list of x and y coordinates of "discrete lines" with
specified numbers of points, forming a quadrilateral with vertices given
by input coordinates}
\usage{
makequad(
  x1 = 4,
  y1 = 4,
  x2 = 10,
  y2 = 10,
  x3 = 11,
  y3 = 14,
  x4 = 15,
  y4 = 0,
  npts = 30,
  make_open = TRUE
)
}
\arguments{
\item{x1}{numeric(1)}

\item{y1}{numeric(1)}

\item{x2}{numeric(1)}

\item{y2}{numeric(1)}

\item{x3}{numeric(1)}

\item{y3}{numeric(1)}

\item{x4}{numeric(1)}

\item{y4}{numeric(1)}

\item{npts}{numeric(1) tells number of points to fill for each side}
}
\description{
produce a list of x and y coordinates of "discrete lines" with
specified numbers of points, forming a quadrilateral with vertices given
by input coordinates
}
\note{
chull is used to ensure sides do not cross
}
\examples{
str(makequad())
}
