% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketri.R
\name{maketri}
\alias{maketri}
\title{produce a list of x and y coordinates of "discrete lines" with
specified numbers of points, forming a triangle with vertices given
by input coordinates}
\usage{
maketri(x1 = 4, y1 = 4, x2 = 10, y2 = 10, x3 = 6, y3 = 14, npts = 20)
}
\arguments{
\item{x1}{numeric(1)}

\item{y1}{numeric(1)}

\item{x2}{numeric(1)}

\item{y2}{numeric(1)}

\item{x3}{numeric(1)}

\item{y3}{numeric(1)}

\item{npts}{numeric(1) tells number of points to fill for each side}
}
\description{
produce a list of x and y coordinates of "discrete lines" with
specified numbers of points, forming a triangle with vertices given
by input coordinates
}
\examples{
str(maketri())
}
