% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getApple.R
\name{mastPath}
\alias{mastPath}
\title{get path to activated mast cell image}
\usage{
mastPath()
}
\description{
get path to activated mast cell image
}
\note{
downloaded from http://cellimagelibrary.org/images/222
}
