% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islr_cnn_code.R
\name{model_probs}
\alias{model_probs}
\title{report estimated model properties for a given ImageArray}
\usage{
model_probs(model, iarr, roundto = 3)
}
\arguments{
\item{iarr}{ImageArray instance}

\item{fitted}{islr_cnn model instance}
}
\value{
vector of fitted class probabilities
}
\description{
report estimated model properties for a given ImageArray
}
\examples{
litsh = make_shape_iarr(nimages=3)
preview(litsh)
smod = load_shape_cnn()
model_probs(smod, litsh)
}
