% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranbox.R
\name{ranbox}
\alias{ranbox}
\title{crude approach to a random rectangle in 32x32 rectangle, can bleed over}
\usage{
ranbox(...)
}
\description{
crude approach to a random rectangle in 32x32 rectangle, can bleed over
}
\examples{
set.seed(2345)
plot(0,0,type="n", xlim=c(1,32), ylim=c(1,32))
for (i in seq_len(30)) { lines(ranbox()) }
}
