% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makequad.R
\name{ranquad}
\alias{ranquad}
\title{produce coordinates of a random "quadrilateral" in 64 x 64 plane}
\usage{
ranquad(npts = 50, side_plane = 64)
}
\arguments{
\item{npts}{numeric(1) passed to `maketri`}

\item{side_plane}{numeric(1) length of square side for rendering surface}
}
\description{
produce coordinates of a random "quadrilateral" in 64 x 64 plane
}
\examples{
set.seed(2345)
plot(0,0,type="n", xlim=c(1,64), ylim=c(1,64))
for (i in seq_len(30)) { lines(ranquad()) }
}
