% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketri.R
\name{rantri}
\alias{rantri}
\title{produce coordinates of a random "triangle" in 64 x 64 plane}
\usage{
rantri(npts = 50, side_plane = 64)
}
\arguments{
\item{npts}{numeric(1) passed to `maketri`}

\item{side_plane}{numeric(1) length of square side used}
}
\description{
produce coordinates of a random "triangle" in 64 x 64 plane
}
\examples{
plot(0,0,type="n",xlim=c(0,1), ylim=c(0,1))
set.seed(2345)
tt = rantri()
graphics::rasterImage(load_jpeg(tt$x, tt$y, siz=sample(c(20,40,60), size=1)), 0, 0, 1, 1, interpolate=FALSE)
}
