% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islr_cnn_code.R
\name{restore_islr_cnn}
\alias{restore_islr_cnn}
\title{restore fitted model and addons}
\usage{
restore_islr_cnn(folder)
}
\arguments{
\item{folder}{character(1) path to a previously saved islr_cnn}
}
\value{
instance of fitted islr_cnn with addons describing package version for
creation
}
\description{
restore fitted model and addons
}
\examples{
restore_islr_cnn(system.file("extdata", "shapemodf", package="littleDeep"))
}
