% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEBImage.R
\name{runSobel}
\alias{runSobel}
\title{perform edge detection via Sobel algorithm}
\usage{
runSobel(iarr1)
}
\arguments{
\item{iarr1}{ImageArray instance}
}
\description{
perform edge detection via Sobel algorithm
}
\examples{
data(ciftrain1k)
ore = runSobel(ciftrain1k[4])
opar = par(no.readonly=TRUE)
par(mfrow=c(2,2), mar=c(1,1,1,1))
plotOne(ciftrain1k[4])
plotOne(ore, transpose=TRUE)
plotOne(ore, transpose=TRUE, interpolate=FALSE)
plotOne(ciftrain1k[4], interpolate=FALSE)
par(opar)
}
