% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islr_cnn_code.R
\name{save_islr_cnn}
\alias{save_islr_cnn}
\title{save a fitted model in a folder with hdf5 and list components}
\usage{
save_islr_cnn(islr_cnn, folder, objname = "model.h5")
}
\arguments{
\item{islr_cnn}{instance of islr_cnn}

\item{folder}{character(1) must not exist}

\item{objname}{character(1) defaults to "model.h5"}
}
\value{
folder and object name as path
}
\description{
save a fitted model in a folder with hdf5 and list components
}
