% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCE2ImageArray.R
\name{SCE2ImageArray}
\alias{SCE2ImageArray}
\title{convert assay data from SCE to ImageArray}
\usage{
SCE2ImageArray(sce, classvec, dims = c(32, 32, 3), vfun = rowSds, ...)
}
\arguments{
\item{sce}{SingleCellExperiment instance}

\item{classvec}{vector of labels}

\item{dims}{dimensions of a single image}

\item{vfun}{a function to be applied to rows of assay(sce) to measure feature-wise 
variation in expression across cells}

\item{\dots}{not used}
}
\description{
convert assay data from SCE to ImageArray
}
