% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islr_cnn_code.R
\name{scores_string}
\alias{scores_string}
\title{make a nice string representation of types and scores for a single image}
\usage{
scores_string(fitted, iarr1, n = 3)
}
\arguments{
\item{fitted}{islr_cnn model instance}

\item{iarr1}{ImageArray with one image}

\item{n}{numeric(1) number of scores to format, default to 3}
}
\description{
make a nice string representation of types and scores for a single image
}
\examples{
litsh = make_shape_iarr(nimages=3)
preview(litsh)
smod = load_shape_cnn()
scores_string(smod, litsh[1])
purp = system.file("jpegs", "newpurple.jpg", package="littleDeep")
c3 = restore_islr_cnn(system.file("extdata", "cif3", package="littleDeep"))
ia = jpeg2iarr(purp, "rose", c3$typelevels)
scores_string(c3$model, ia)
}
