% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageArray.R
\name{show,ImageArray-method}
\alias{show,ImageArray-method}
\title{display information about ImageArray instance}
\usage{
\S4method{show}{ImageArray}(object)
}
\description{
display information about ImageArray instance
}
