% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_img.R
\name{show_jpg}
\alias{show_jpg}
\title{produce raster image of a jpg file}
\usage{
show_jpg(fn, main = "jpeg", ...)
}
\arguments{
\item{fn}{character(1) path to jpg}

\item{main}{character(1) title defaults to 'jpeg'}

\item{\dots}{passed to raster::rasterImage}
}
\value{
rasterImage is called
}
\description{
produce raster image of a jpg file
}
