% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_img.R
\name{show_jpg2}
\alias{show_jpg2}
\title{a magick-based solution to presenting jpg data, includes rotation capability}
\usage{
show_jpg2(fn, main = "jpeg", degrees = 0, ...)
}
\arguments{
\item{fn}{character(1) path to jpg}

\item{main}{character(1) title defaults to 'jpeg'}
}
\description{
a magick-based solution to presenting jpg data, includes rotation capability
}
\examples{
apl = system.file("jpegs", "oneBigApple.jpg", package="littleDeep")
show_jpg(apl)
}
