% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageArray.R
\name{[,ImageArray,numeric,missing,missing-method}
\alias{[,ImageArray,numeric,missing,missing-method}
\title{simple extraction with bracket}
\usage{
\S4method{[}{ImageArray,numeric,missing,missing}(x, i, j, drop)
}
\description{
simple extraction with bracket
}
