% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress_keras_warning.R
\name{suppress_keras_warnings}
\alias{suppress_keras_warnings}
\title{from Claude, to help reduce noise in Apr 2026 usage}
\usage{
suppress_keras_warnings()
}
\description{
from Claude, to help reduce noise in Apr 2026 usage
}
