% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docifar100.R
\name{train_cifar100}
\alias{train_cifar100}
\title{build islr_cnn instance trained on cifar100}
\usage{
train_cifar100(saveto = tempfile(), nEpochs = 25, ...)
}
\arguments{
\item{saveto}{character(1) passed to save_islr_cnn, folder where hdf5 and additional metadata are stored}

\item{nEpochs}{numeric(1) used in `islr_cnn`}

\item{\dots}{passed to `islr_cnn`}
}
\description{
build islr_cnn instance trained on cifar100
}
