% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageArray.R
\name{ImageArray}
\alias{ImageArray}
\title{constructor for ImageArray}
\usage{
ImageArray(arr, types, typelevels)
}
\arguments{
\item{arr}{4D array}

\item{types}{character()}

\item{typelevels}{optional character()
in case ImageArray is a limited collection of images from a larger family 
of possible types}
}
\description{
constructor for ImageArray
}
\examples{
  cifmeta = jsonlite::fromJSON(system.file("extdata", "cif.json", package="littleDeep"))
  labs = cifmeta[[1]]$features$fine_label$names
  cifdata = dataset_cifar100()
  ciftrain = cifdata[[1]]$x
  ciflabels = labs[cifdata[[1]]$y+1]
  ciftrainArr = ImageArray(ciftrain, ciflabels)
  preview(ciftrainArr[1:9])
}
