% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageArray.R
\name{ImageArray}
\alias{ImageArray}
\title{constructor for ImageArray}
\usage{
ImageArray(arr, types)
}
\arguments{
\item{arr}{4D array}

\item{types}{character()}
}
\description{
constructor for ImageArray
}
