% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEBImage.R
\name{plotOne}
\alias{plotOne}
\title{plot a length-1 ImageArray using as.raster}
\usage{
plotOne(iarr1, transpose = FALSE, ...)
}
\arguments{
\item{iarr1}{ImageArray instance}

\item{transpose}{logical(1) use aperm(..., c(2,1,3)) before rendering}
}
\description{
plot a length-1 ImageArray using as.raster
}
\examples{
data(ciftrain1k)
ore = runSobel(ciftrain1k[4])
opar = par(no.readonly=TRUE)
par(mfrow=c(2,2),mar=c(1,1,1,1))
plotOne(ciftrain1k[4])
plotOne(ore, transpose=TRUE)
plotOne(ore, transpose=TRUE, interpolate=FALSE)
plotOne(ciftrain1k[4], interpolate=FALSE)
par(opar)
}
