% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageArray.R
\name{preview}
\alias{preview}
\title{produce labeled 3x3 array of images via plot.raster}
\usage{
preview(iarr, ...)
}
\arguments{
\item{iarr}{ImageArray instance}

\item{\dots}{passed to plot.raster and thence to rasterImage; `interpolate=TRUE` can be useful}
}
\description{
produce labeled 3x3 array of images via plot.raster
}
