% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cifar100.R
\name{run_cifar100}
\alias{run_cifar100}
\title{run the ISLR CNN for CIFAR100 data}
\usage{
run_cifar100()
}
\value{
a list with elements history and model
}
\description{
run the ISLR CNN for CIFAR100 data
}
\note{
Takes 15+ minutes on a CPU
}
